Starting LCA and DAG afresh
Once I began the new assignment, to facilitate the program to work with directed acyclic graphs, I found it extremely awkward to update my code to allow for this.
I am also a lot more comfortable working with GitHub now, and after coming to grasps with it and this new assignment, I have decided to start afresh.

I am renaming my original code as OldLCA, so that my old work is still visible, but it is not the program with which I will continue to work.